Remove empty token variable
This action was copied from `vshn/appcat`, where the component_token
variable is defined and used for pulling. However it doesn't exist
in the component-appcat repository. Leading to an error during release.

By removing it, it will default to the actions token, which has enough
permissions to do the release on this repository.